renamed server module to protobuf
