+++ src/api/reveaal_context.rs
@@ -1,6 +1,6 @@
-use crate::api::server::server::ecdar_backend_client::EcdarBackendClient;
-use crate::api::server::server::ecdar_backend_server::EcdarBackend;
-use crate::api::server::server::{
+use crate::api::server::protobuf::ecdar_backend_client::EcdarBackendClient;
+use crate::api::server::protobuf::ecdar_backend_server::EcdarBackend;
+use crate::api::server::protobuf::{
     QueryRequest, QueryResponse, SimulationStartRequest, SimulationStepRequest,
     SimulationStepResponse, UserTokenResponse,
 };
