+++ src/tests/api/access_logic.rs
@@ -1,5 +1,5 @@
-use crate::api::server::server::ecdar_api_server::EcdarApi;
-use crate::api::server::server::{CreateAccessRequest, DeleteAccessRequest, UpdateAccessRequest};
+use crate::api::server::protobuf::ecdar_api_server::EcdarApi;
+use crate::api::server::protobuf::{CreateAccessRequest, DeleteAccessRequest, UpdateAccessRequest};
 use crate::entities::access;
 use crate::tests::api::helpers::{get_mock_concrete_ecdar_api, get_mock_services};
 use mockall::predicate;
