+++ src/tests/api/helpers.rs
@@ -4,9 +4,9 @@ use crate::api::auth::TokenType;
 use crate::api::context_collection::ContextCollection;
 use crate::api::ecdar_api::ConcreteEcdarApi;
 use crate::api::hashing_context::HashingContextTrait;
-use crate::api::server::server::ecdar_backend_server::EcdarBackend;
-use crate::api::server::server::ModelInfo;
-use crate::api::server::server::{
+use crate::api::server::protobuf::ecdar_backend_server::EcdarBackend;
+use crate::api::server::protobuf::ModelInfo;
+use crate::api::server::protobuf::{
     QueryRequest, QueryResponse, SimulationStartRequest, SimulationStepRequest,
     SimulationStepResponse, UserTokenResponse,
 };
